+++ docs/Changelog.md
@@ -4,6 +4,7 @@
 ## v3.3 (upcoming)
 - FIX: recognize JUnit file format which wrote as single line ([JENKINS-45723](https://issues.jenkins-ci.org/browse/JENKINS-45723))
 - FIX: dependency that require code v2.62+
+- FIX: show only chart legend ([JENKINS-45539](https://issues.jenkins-ci.org/browse/JENKINS-45539))
 - FIX: virtualenv error in job which contains spaces in name
 - FIX: publisher for more than one sourceDataFiles ([JENKINS-46046](https://issues.jenkins-ci.org/browse/JENKINS-46046)). Contributed by [Till Neunast](https://github.com/tilln)
 - FIX: logging for expert criteria works with pipeline without throwing exceptions. Contributed by [Till Neunast](https://github.com/tilln)